requirement.txt
xlrd
python-docx
xlwt
numpy
pywin32
pdfminer3k
pandas
jieba